requirements.txt
requests
python-dotenv
PyMuPDF
pypdf
langgraph
langsmith
langchain_anthropic
langchain-google-genai
langchain
langchain_core
langchain_openai 
chromadb
langchain_community
langchain-chroma
langchain-google-vertexai
pandas
bs4
streamlit
loguru
pytesseract